pyproject.toml
[tool.poetry]
name = "demoqa-diplome"
version = "0.1.0"
description = ""
authors = ["Vitaliy Zakharov <zak125@mail.ru>"]
readme = "README.md"
packages = [{include = "demoqa_diplome"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = ">=7.4.3,<7.5.0"
allure-pytest = "^2.13.2"
python-dotenv = ">=1.0.0,<1.1.0"
selene = "2.0.0rc4"
selenium = ">=4.14.0,<4.15.0"
jsonschema = "^4.20.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
